Refactored include additional query processing class into new utils file to allow for easier usability.  Moved created getter method in serializer to work with both node list and node detail.
