+++ api/nodes/serializers.py
@@ -1,3 +1,4 @@
+from .utils import IncludeAdditionalQuery
 from rest_framework import serializers as ser
 
 from website.models import Node
@@ -61,7 +62,7 @@ class NodeSerializer(JSONAPISerializer):
                                                             'node have implicit read permissions for all child nodes',
                               )
 
-    additional_query_params = ser.DictField(required=False, read_only=True)
+    additional_query_params = ser.SerializerMethodField(required=False, read_only=True)
 
     # TODO: finish me
 
@@ -96,6 +97,11 @@ class NodeSerializer(JSONAPISerializer):
 
     def get_pointers_count(self, obj):
         return len(obj.nodes_pointer)
+
+    def get_additional_query_params(self, obj):
+        include = IncludeAdditionalQuery(obj, self.context['request'])
+        ret = include.get_additional_query()
+        return ret
 
     @staticmethod
     def get_properties(obj):
