+++ api/nodes/utils.py
@@ -1,0 +1,56 @@
+from rest_framework.exceptions import NotFound
+from framework.auth.core import Auth
+
+
+class IncludeAdditionalQuery(object):
+
+    def __init__(self, obj, request):
+        self.obj = obj
+        self.request = request
+
+    def get_additional_query(self):
+        query = {}
+        if 'include' in self.request.query_params:
+            params = self.request.query_params['include'].split(',')
+            if 'children' in params:
+                query['children'] = self.get_children()
+                params.remove('children')
+            if 'contributors' in params:
+                query['contributors'] = self.get_contributors()
+                params.remove('contributors')
+            if 'pointers' in params:
+                query['pointers'] = self.get_pointers()
+                params.remove('pointers')
+            if params != []:
+                params_string = ', '.join(params)
+                raise NotFound('The following arguments cannot be found: {}'.format(params_string))
+        return query
+
+    def get_children(self):
+        nodes = {}
+        for node in self.obj.nodes:
+            if node.can_view(Auth(self.request.user)) and node.primary:
+                nodes[node._id] = {
+                    'title': node.title,
+                    'description': node.description,
+                    'is_public': node.is_public
+                }
+        return nodes
+
+    def get_contributors(self):
+        contributors = {}
+        for contributor in self.obj.contributors:
+            contributors[contributor._id] = {
+                'username': contributor.username,
+                'bibliographic': self.obj.get_visible(contributor),
+                'permissions': self.obj.get_permissions(contributor)
+            }
+        return contributors
+
+    def get_pointers(self):
+        pointers = {}
+        for pointer in self.obj.nodes_pointer:
+            pointers[pointer._id] = {
+                'title': pointer.title,
+            }
+        return pointers
